Refactor : change package structure
