+++ src/main/java/ladder/view/InputView.java
@@ -1,6 +1,9 @@
 package ladder.view;
 
-import ladder.model.*;
+import ladder.model.player.Player;
+import ladder.model.player.Players;
+import ladder.model.prize.LadderPrize;
+import ladder.model.row.Height;
 import ladder.util.ScannerUtil;
 
 import java.util.Arrays;
