+++ src/test/java/ladder/domain/prize/LadderPrizeTest.java
@@ -1,6 +1,6 @@
-package ladder;
+package ladder.domain.prize;
 
-import ladder.model.LadderPrize;
+import ladder.model.prize.LadderPrize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
